pyproject.toml
[project]
name = "tripy"
version = "0.1.0"
description = ""
authors = [{name = "Parth Chadha", email="pchadha@nvidia.com"}]
readme = "README.md"
dependencies = [
  "mlir-tensorrt-compiler==0.1.25+cuda12.trt100",
  "mlir-tensorrt-runtime==0.1.25+cuda12.trt100",
]

[project.optional-dependencies]
dev = [
  "colored==2.2.3",
  "jinja2==3.1.2",
  "numpy==1.23.5",
  "cupy-cuda12x",
  "pre-commit==3.6.0",
]
docs = [
  "sphinx==7.2.6",
  "sphinx-nefertiti==0.2.1",
  "sphinx-copybutton==0.5.2",
  "sphinx-toolbox==3.5.0",
  "docutils==0.20.1",
  "myst-parser==2.0.0",
  "sphinxcontrib-mermaid==0.9.2",
]
test = [
  "torch==2.0.0",
  "pytest==7.1.3",
  "pytest-virtualenv==1.7.0",
  "pytest-cov==4.1.0",
  "jax[cuda12_local]",
  "jaxlib==0.4.23+cuda12.cudnn89",
  "coverage==7.4.1",
  "vulture==2.11",
]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
)/
'''

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
addopts = "--strict-markers"
markers = [
    "l0: Marks tests to run on each merge request.",
    "l1: Marks tests to run at a nightly cadence.",
]
